// packages/components/src/rating/snice-rating.types.ts
export type RatingPrecision = 'full' | 'half';
export type RatingSize = 'small' | 'medium' | 'large';

export interface SniceRatingElement extends HTMLElement {
  value: number;
  max: number;
  icon: string;
  emptyIcon: string;
  size: RatingSize;
  readonly: boolean;
  precision: RatingPrecision;
}

export interface SniceRatingEventMap {
  'rating-change': CustomEvent<{ value: number }>;
}